+++ packages/headless/src/features/triggers/trigger-analytics-actions.ts
@@ -45,3 +45,19 @@ export const logTriggerRedirect = makeAnalyticsAction(
     return;
   }
 );
+
+/**
+ * Log trigger execute
+ */
+export const logTriggerExecute = makeAnalyticsAction(
+  'analytics/trigger/execute',
+  AnalyticsType.Search,
+  (client, state) => {
+    if (state.triggers?.execute) {
+      return client.logTriggerExecute({
+        executed: state.triggers.execute.functionName,
+      });
+    }
+    return;
+  }
+);