+++ packages/headless/src/features/triggers/triggers-state.ts
@@ -1,3 +1,5 @@
+import {ExecuteTriggerParams} from '../../api/search/trigger';
+
 export interface TriggerState {
   /**
    * The URL to redirect the user to after receiving a redirection trigger.
@@ -10,6 +12,11 @@ export interface TriggerState {
   query: string;
 
   /**
+   * The JavaScript function to be executed after receiving an execute trigger.
+   */
+  execute: {functionName: string; params: ExecuteTriggerParams};
+
+  /**
    * The notification to display to the user after receiving a notification trigger.
    */
   notification: string;
@@ -18,5 +25,6 @@ export interface TriggerState {
 export const getTriggerInitialState: () => TriggerState = () => ({
   redirectTo: '',
   query: '',
+  execute: {functionName: '', params: []},
   notification: '',
 });