+++ packages/quantic/force-app/main/default/lwc/quanticSearchBox/quanticSearchBox.js
@@ -77,6 +77,9 @@ export default class QuanticSearchBox extends LightningElement {
   }
 
   updateState() {
+    if(this.state.value !== this.searchBox.state.value) {
+      this.updateSearchboxText(this.searchBox.state.value);
+    }
     this.state = this.searchBox.state;
   }
 
@@ -93,13 +96,16 @@ export default class QuanticSearchBox extends LightningElement {
   }
 
   setHighlighted() {
-    const options = this.getSuggestionElements();
-    const option = options[this.selectionIndex];
+    const suggestions = this.getSuggestionElements();
+    if (!suggestions.length) {
+      return;
+    }
+    const suggestion = suggestions[this.selectionIndex];
 
     this.resetHighlighted();
-    option.setAttribute('aria-selected', 'true');
-    option.classList.add('slds-has-focus');
-    this.input.value = option.textContent;
+    suggestion.setAttribute('aria-selected', 'true');
+    suggestion.classList.add('slds-has-focus');
+    this.input.value = suggestion.textContent;
   }
 
   resetHighlighted() {